[REF] mail: move mail.message/mark_as_read to new service
Part of task-3265211
